backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular[sidecar]>=0.27
psycopg[binary]>=3.1
Pillow>=10.0
django-filter>=23.5
